requirements.txt
pandas>=2.0.0
streamlit>=1.30.0
pytest>=7.4.0
openai>=1.30.0
pydantic>=2.5.0
chromadb>=0.4.22
python-dotenv>=1.0.0
rich>=13.7.0
tiktoken>=0.5.0
